feat: Phase 2 — RSC化 + 公開ビュー（認証ゲート撤廃・unstable_cache・generateMetadata）
## 変更概要

### 公開ビュー化
- ルートを `/viewer-auth` → `/view` リダイレクトに変更
- `/api/sheets` および `/api/sheets/content` の HMAC セッション認証ゲートを撤廃
  （閲覧は完全公開。編集/管理認証は Phase 3 Better Auth で実装予定）
- compare-client の `AuthError` → `/viewer-auth` リダイレクトを除去

### RSC 化 + サーバーサイドキャッシュ
- `app/view/page.tsx` を RSC に昇格。`unstable_cache(revalidate:3600)` でシート一覧を
  サーバーキャッシュ（ビルド結果: `Revalidate: 1h`）
- インタラクティブUI（検索・比較チェックボックス）は `sheets-list-client.tsx` に分離
- `app/view/[path]/page.tsx` を RSC に昇格。`generateMetadata` で SEO/OG タイトルを出力
  PDF ダウンロード等の client 処理は `sheet-view-client.tsx` に分離

### 'use client' の明示化
- `skill-sheet-viewer.tsx` に `'use client'` を追加（RSC から参照される可能性への備え）

## 検証
- `pnpm -r type-check` ✅
- `next build` ✅（65テスト全グリーン）
- `/view` → `Revalidate: 1h` の静的プリレンダリング確認
