+++ apps/web/app/api/sheets/content/route.ts
@@ -1,16 +1,11 @@
 import { NextRequest, NextResponse } from 'next/server';
 
 import { fetchSheetFile, isValidSheetPath } from '@/server/github-sheets';
-import { SESSION_COOKIE_NAME, verifySessionToken } from '@/server/session';
 
 export const runtime = 'nodejs';
 export const dynamic = 'force-dynamic';
 
 export async function GET(req: NextRequest) {
-  if (!verifySessionToken(req.cookies.get(SESSION_COOKIE_NAME)?.value)) {
-    return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
-  }
-
   const path = req.nextUrl.searchParams.get('path');
   if (!path) {
     return NextResponse.json({ error: 'Missing path parameter' }, { status: 400 });
@@ -22,7 +17,7 @@ export async function GET(req: NextRequest) {
   try {
     const content = await fetchSheetFile(path);
     return NextResponse.json(content, {
-      headers: { 'Cache-Control': 'private, no-store' },
+      headers: { 'Cache-Control': 'public, s-maxage=3600, stale-while-revalidate=86400' },
     });
   } catch (err) {
     console.error('GET /api/sheets/content failed:', err);
