+++ apps/web/app/api/sheets/route.ts
@@ -1,20 +1,15 @@
-import { NextRequest, NextResponse } from 'next/server';
+import { NextResponse } from 'next/server';
 
 import { listSheets } from '@/server/github-sheets';
-import { SESSION_COOKIE_NAME, verifySessionToken } from '@/server/session';
 
 export const runtime = 'nodejs';
 export const dynamic = 'force-dynamic';
 
-export async function GET(req: NextRequest) {
-  if (!verifySessionToken(req.cookies.get(SESSION_COOKIE_NAME)?.value)) {
-    return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
-  }
-
+export async function GET() {
   try {
     const sheets = await listSheets();
     return NextResponse.json(sheets, {
-      headers: { 'Cache-Control': 'private, no-store' },
+      headers: { 'Cache-Control': 'public, s-maxage=3600, stale-while-revalidate=86400' },
     });
   } catch (err) {
     console.error('GET /api/sheets failed:', err);
