+++ apps/web/app/view/[path]/page.tsx
@@ -1,117 +1,44 @@
-'use client';
+import { type Metadata } from 'next';
+import { unstable_cache } from 'next/cache';
+import { notFound } from 'next/navigation';
 
-import { useParams, useRouter } from 'next/navigation';
-import { useEffect, useState } from 'react';
+import { fetchSheetFile, isValidSheetPath } from '@/server/github-sheets';
 
-import { Loader2 } from 'lucide-react';
-import { toast } from 'sonner';
+import SheetViewClient from './sheet-view-client';
 
-import Header from '@/component/header';
-import SkillSheetViewer from '@/component/skill-sheet-viewer';
-import { AuthError, fetchSheet } from '@/lib/skillsheet-client';
-
-interface SkillSheet {
-  title: string;
-  content: string;
+interface PageProps {
+  params: Promise<{ path: string }>;
 }
 
-const REVOKE_OBJECT_URL_DELAY_MS = 100;
+const getCachedSheet = unstable_cache(
+  async (path: string) => fetchSheetFile(path),
+  ['sheet'],
+  { tags: ['sheets'], revalidate: 3600 },
+);
 
-const SheetViewPage = () => {
-  const router = useRouter();
-  const params = useParams<{ path?: string | string[] }>();
-  const rawPath = params.path;
-  const path = Array.isArray(rawPath) ? rawPath[0] : rawPath;
-  const [skillSheet, setSkillSheet] = useState<SkillSheet | null>(null);
-  const [loading, setLoading] = useState(true);
-  const [error, setError] = useState<string | null>(null);
-  const [pdfLoading, setPdfLoading] = useState(false);
-
-  useEffect(() => {
-    if (!path) {
-      router.push('/view');
-      return;
-    }
-
-    const loadSkillSheet = async () => {
-      try {
-        setLoading(true);
-        const data = await fetchSheet(path);
-        setSkillSheet({ title: data.title, content: data.content });
-        setError(null);
-      } catch (err) {
-        if (err instanceof AuthError) {
-          router.push('/viewer-auth');
-          return;
-        }
-        console.error('Error fetching skill sheet:', err);
-        setError('スキルシートの読み込みに失敗しました。');
-      } finally {
-        setLoading(false);
-      }
+export async function generateMetadata({ params }: PageProps): Promise<Metadata> {
+  const { path } = await params;
+  const decodedPath = decodeURIComponent(path);
+  if (!isValidSheetPath(decodedPath)) return {};
+  try {
+    const sheet = await getCachedSheet(decodedPath);
+    return {
+      title: `${sheet.title} | エンジニアスキルシート`,
+      openGraph: { title: sheet.title, type: 'profile' },
     };
-
-    void loadSkillSheet();
-  }, [router, path]);
-
-  const handleDownloadPdf = async () => {
-    if (!skillSheet) return;
-
-    try {
-      setPdfLoading(true);
-
-      const [{ pdf }, { SkillSheetPDF }] = await Promise.all([
-        import('@react-pdf/renderer'),
-        import('@/component/pdf-export'),
-      ]);
-
-      const blob = await pdf(<SkillSheetPDF title={skillSheet.title} content={skillSheet.content} />).toBlob();
-
-      const url = URL.createObjectURL(blob);
-      const link = document.createElement('a');
-      link.href = url;
-      link.download = `${skillSheet.title}.pdf`;
-      document.body.appendChild(link);
-      link.click();
-      document.body.removeChild(link);
-      setTimeout(() => {
-        URL.revokeObjectURL(url);
-      }, REVOKE_OBJECT_URL_DELAY_MS);
-
-      toast.success('PDFをダウンロードしました');
-    } catch (err) {
-      console.error('Error generating PDF:', err);
-      toast.error('PDFの生成に失敗しました');
-    } finally {
-      setPdfLoading(false);
-    }
-  };
-
-  if (loading) {
-    return (
-      <div className="flex min-h-screen items-center justify-center">
-        <Loader2 className="size-10 animate-spin text-primary" />
-      </div>
-    );
+  } catch {
+    return {};
   }
+}
 
-  if (error) {
-    return (
-      <div className="flex min-h-screen flex-col items-center justify-center gap-2 px-4 text-center">
-        <h2 className="text-2xl font-bold">エラー</h2>
-        <p className="text-muted-foreground">{error}</p>
-      </div>
-    );
+export default async function SheetViewPage({ params }: PageProps) {
+  const { path } = await params;
+  const decodedPath = decodeURIComponent(path);
+  if (!isValidSheetPath(decodedPath)) notFound();
+  try {
+    const sheet = await getCachedSheet(decodedPath);
+    return <SheetViewClient title={sheet.title} content={sheet.content} />;
+  } catch {
+    notFound();
   }
-
-  if (!skillSheet) return null;
-
-  return (
-    <div>
-      <Header onDownloadPdf={handleDownloadPdf} pdfLoading={pdfLoading} />
-      <SkillSheetViewer skillSheet={skillSheet} />
-    </div>
-  );
-};
-
-export default SheetViewPage;
+}
