+++ apps/web/src/component/skill-sheet-viewer.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { useEffect, useRef, useState } from 'react';
 import ReactMarkdown from 'react-markdown';
